refactor move all tado api structs into own file
